refactor: remove redundant comments and streamline dashboard access logic by redirecting unauthenticated users to sign-in
