+++ src/app/auth/sign-up/page.tsx
@@ -3,7 +3,6 @@ import { getCurrentUser } from '@/src/shared/lib/auth'
 import { redirect } from 'next/navigation'
 
 export default async function Page() {
-  // 이미 로그인한 사용자는 대시보드로 리디렉션
   const user = await getCurrentUser()
   if (user) {
     redirect('/dashboard')
